requirements.txt
transformers==4.25.1
pytorch_lightning==1.3.5
torchmetrics==0.5.0
nltk==3.6.1
py-rouge
pandas
six
